+++ emav/dofft.c
@@ -60,11 +60,11 @@ alloc_space(int create)
 		free(sbuf);
 	    }
 	    size = 0;
-	    abuf = (float *) calloc((n + 2), sizeof(float));
-	    bbuf = (float *) calloc((n + 2), sizeof(float));
-	    fbuf = (float *) calloc((n + 2), sizeof(float));
-	    phbf = (float *) calloc((n + 2), sizeof(float));
-	    sbuf = (float *) calloc((n + 2), sizeof(float));
+	    abuf = (unsigned long)((float *) calloc((n + 2), sizeof(float)));
+	    bbuf = (unsigned long)((float *) calloc((n + 2), sizeof(float)));
+	    fbuf = (unsigned long)((float *) calloc((n + 2), sizeof(float)));
+	    phbf = (unsigned long)((float *) calloc((n + 2), sizeof(float)));
+	    sbuf = (unsigned long)((float *) calloc((n + 2), sizeof(float)));
 	    if (fbuf != NULL)
 		size = n;
 	}
@@ -240,22 +240,22 @@ show_fft(int xrange, int yrange, int levref, int color, int reset)
 
     if (xrange < 3) {
 	n = size;
-	dt = ((float) n) / rate;
+	dt = (float)(((float) n) / rate);
 	lpsval = lpsflg ? 10 * log10(dt) : 0;
 	for (i = 0; i < n; i++) {
 	    a = fbuf[i];
 	    a = (i == 0 || a < EMIN) ? EMIN : a;
-	    ibuf[i] = (int) (10 * log10(a) + lpsval + 0.5);
+	    ibuf[i] = (short)((int) (10 * log10(a) + lpsval + 0.5));
 	}
     } else {
 	n = size / 2;
-	dt = ((float) n) / rate;
+	dt = (float)(((float) n) / rate);
 	lpsval = lpsflg ? 10 * log10(dt) : 0;
 	for (i = 0; i < n; i++) {
 	    a = fbuf[i++];
 	    a += fbuf[i];
 	    a = (i == 1 || a < EMIN) ? EMIN : a;
-	    ibuf[i / 2] = (int) (10 * log10(a) + lpsval + 0.5);
+	    ibuf[i / 2] = (short)((int) (10 * log10(a) + lpsval + 0.5));
 	}
     }
     w = w_dis;
@@ -388,9 +388,9 @@ cal_gain(SAV_FFT *pft, int npts)
             free(pft->rp);
         if (pft->ph != NULL)
             free(pft->ph);
-	pft->rp = (float *) malloc(npts * sizeof(float));
-	pft->ph = (float *) malloc(npts * sizeof(float));
-	pft->npts = npts;
+	pft->rp = (unsigned long)((float *) malloc(npts * sizeof(float)));
+	pft->ph = (unsigned long)((float *) malloc(npts * sizeof(float)));
+	pft->npts = (short)(npts);
     }
     memcpy(pft->rp, sbuf, npts * sizeof(float));
     memcpy(pft->ph, phbf, npts * sizeof(float));
@@ -595,8 +595,8 @@ get_level(double fr, int nsb, int typ, double df, float *slv, float *nlv, float 
 	j = nint((df * size) / rate);
     }
     n = 1 + 2 * ns;
-    asp = (float *) calloc(n, 2 * sizeof(float));
-    bsp = (float *) calloc(n, 2 * sizeof(float));
+    asp = (unsigned long)((float *) calloc(n, 2 * sizeof(float)));
+    bsp = (unsigned long)((float *) calloc(n, 2 * sizeof(float)));
     fetch_spec(fr, ns, asp, bsp);
 
 /* remove correlated noise to reduce artifacts */
@@ -685,8 +685,8 @@ check_noise(double f, double df, double nf_max_dif, short nfsb)
     ns = nfsb;
     j = nint((df * size) / rate);
     n = 1 + 2 * ns;
-    asp = (float *) calloc(n, 2 * sizeof(float));
-    bsp = (float *) calloc(n, 2 * sizeof(float));
+    asp = (unsigned long)((float *) calloc(n, 2 * sizeof(float)));
+    bsp = (unsigned long)((float *) calloc(n, 2 * sizeof(float)));
     fetch_spec(f, ns, asp, bsp);
 
 /* compute noise level */
@@ -743,7 +743,7 @@ mp_transfer_write(char *fn)
 	fprintf(fp, ";    f     mg     ph\n");
         tpi = 2 * M_PI;
 	for (i = 1; i <mps_nf; i++) {
-	    fr = mps_df * i;
+	    fr = (float)(mps_df * i);
 	    mg = mps_mg[i];
 	    ph = mps_ph[i] / tpi;	// convert rad to cyc
 	    fprintf(fp, " %6.0f %7.4f %6.3f\n", fr, mg, ph);
@@ -761,12 +761,12 @@ mp_transfer_set(int nf, float *fr, float *mg, float *ph)
     free_zero(mps_mg);
     mps_nf = buflen / 2;
     mps_df = rate / (float) buflen; // df=Hz
-    mps_fm = mps_df * mps_nf;
-    mps_mg = (float *) calloc(mps_nf * 2, sizeof(float));
+    mps_fm = (float)(mps_df * mps_nf);
+    mps_mg = (unsigned long)((float *) calloc(mps_nf * 2, sizeof(float)));
     mps_ph = mps_mg + mps_nf;
     j = 0;
     for (i = 0; i < mps_nf; i++) {
-	f = i * mps_df;
+	f = (float)(i * mps_df);
 	if (nf < 1) {
 	    mps_mg[i] = 0;
 	    mps_ph[i] = 0;
@@ -796,7 +796,7 @@ mp_transfer_mat(char *fn)
 	fd = mptrans_open(fn, &n);
 	if (fd > 0 && n > 0) {
 	    nf = n / 2 + 1;
-	    fr = (float *) calloc(nf * 3, sizeof(float));
+	    fr = (unsigned long)((float *) calloc(nf * 3, sizeof(float)));
 	    mg = fr + nf;
 	    ph = mg + nf;
 	    mptrans_read(fd, fr, mg, ph);
@@ -823,7 +823,7 @@ mp_transfer_txt(char *fn)
 	}
 	fclose(fp);
 	if (nf > 0) {
-	    fr = (float *) calloc(nf * 3, sizeof(float));
+	    fr = (unsigned long)((float *) calloc(nf * 3, sizeof(float)));
 	    mg = fr + nf;
 	    ph = mg + nf;
 	    tpi = (float) (2 * M_PI);
